gateway: begin splitting of mega-file
This proceeds by trying to first move the wrappers at the bottom
of the file (instruments) to their own file. Doing this requires
then extracting some of the first elements like base classes out
to another file. The global variable `KNOWN_WRAPPERS` must be put
in yet another file since it is used in all the other files.

New line counts:

```
    8664 src/omero/gateway/__init__.py
    1630 src/omero/gateway/_core.py
     619 src/omero/gateway/_instruments.py
       1 src/omero/gateway/_wrappers.py
   10828 src/omero/gateway/old.py
```
