+++ src/omero/gateway/_wrappers.py
@@ -1,0 +1,1 @@
+KNOWN_WRAPPERS = {}
